Enemy Speed Reduction fix
Casting the health as floats before the division in the formula

+++ Assets/Scripts/Enemy.cs
@@ -112,7 +112,7 @@ public class Enemy : MonoBehaviour
 
     protected void ReduceMaxSpeed()
     {
-        agent.speed = baseMaxSpeed * (myHealth.CurrentHealth / myHealth.MaxHealth);
+        agent.speed = baseMaxSpeed * ((float)myHealth.CurrentHealth / (float)myHealth.MaxHealth);
     }
 
     public void TakeHit(int damage)
